Adding in outline and base files
In this commit I have added in a basic outline of the final topic model and created corresponding DITA files. While I anticipate having to add more and expand while working, I completed this work in Main to create a solid foundation to work from. The rest of the edits and updates will be completed in corresponding branches.